chore(worker): clean ruff findings + re-verify code pre-prod
Worker linting went from 40 findings to 0 with no behavior change:
- E402: hoist ipaddress/socket/httpx imports to top of runner.py
- B023: bind arc/idx as lambda defaults in deep-vision closure
- RUF046: drop redundant int(round(float)) -> round(...)
- F401/RUF100: remove unused imports and dead noqa directives
- pairwise over zip(words, words[1:]), unpacking over list concat
- rename ambiguous var l -> excerpt_len, wrap >100c lines
- per-file E501 ignore for prompts.py (long prompt templates)

Verified: ruff (api+worker) clean, compileall OK, pytest 6/6,
web typecheck + build green. External API keys, VPS and DNS remain
out of code scope (see handoff section 6/11).

+++ apps/worker/app/pipeline/video_map.py
@@ -6,10 +6,9 @@ from pathlib import Path
 import structlog
 
 from ..models import Transcript, VideoEvent, VideoMap
+from ..prompts import VIDEO_MAP_SYSTEM_PROMPT, video_map_user_prompt
 from ..providers import LLMProvider, ProviderError
 from ..providers.base import ImageInput
-from ..prompts import VIDEO_MAP_SYSTEM_PROMPT, video_map_user_prompt
-from ..settings import get_settings
 from .ffmpeg import FFmpegError, detect_scene_changes, extract_frame
 
 log = structlog.get_logger()
@@ -213,7 +212,9 @@ async def _call_video_map(
                     objects=[str(x) for x in (raw.get("objects") or []) if x],
                     action=str(raw.get("action", "")),
                     transcript_summary=str(raw.get("transcript_summary", ""))[:300],
-                    visual_importance=max(0, min(100, int(round(float(raw.get("visual_importance", 0)))))),
+                    visual_importance=max(
+                        0, min(100, round(float(raw.get("visual_importance", 0))))
+                    ),
                     narrative_role=str(raw.get("narrative_role", "neutral")),
                 )
             )
